Finalize ACD Amortization and Bonds Simulation

+++ Umoa.Titres.Interest.Simulator.Web.Api/Dto/AmortizationLineDto.cs
@@ -3,7 +3,7 @@ namespace Umoa.Titres.Interest.Simulator.Web.Api.Dto;
 
 public class AmortizationLineDto
 {
-    public int Fraction { get; set; }
+    public double Fraction { get; set; }
     public string Date { get; set; }
     public double Encours { get; set; }
     public double Interets { get; set; }
